+++ internal/component/prometheus/exporter/kafka/kafka.go
@@ -8,8 +8,8 @@ import (
 	"github.com/grafana/agent/internal/component/discovery"
 	"github.com/grafana/agent/internal/component/prometheus/exporter"
 	"github.com/grafana/agent/internal/featuregate"
-	"github.com/grafana/agent/internal/static/integrations"
-	"github.com/grafana/agent/internal/static/integrations/kafka_exporter"
+	"github.com/grafana/agent/static/integrations"
+	"github.com/grafana/agent/static/integrations/kafka_exporter"
 	"github.com/grafana/river/rivertypes"
 	"github.com/prometheus/common/config"
 )
@@ -21,8 +21,12 @@ var DefaultArguments = Arguments{
 	AllowConcurrent:         true,
 	MaxOffsets:              1000,
 	PruneIntervalSeconds:    30,
+	OffsetShowAll:           true,
+	TopicWorkers:            100,
 	TopicsFilter:            ".*",
+	TopicsExclude:           "^$",
 	GroupFilter:             ".*",
+	GroupExclude:            "^$",
 }
 
 type Arguments struct {
@@ -33,7 +37,9 @@ type Arguments struct {
 	SASLUsername            string            `river:"sasl_username,attr,optional"`
 	SASLPassword            rivertypes.Secret `river:"sasl_password,attr,optional"`
 	SASLMechanism           string            `river:"sasl_mechanism,attr,optional"`
+	SASLDisablePAFXFast     bool              `river:"sasl_disable_pafx_fast,attr,optional"`
 	UseTLS                  bool              `river:"use_tls,attr,optional"`
+	TlsServerName           string            `river:"tls_server_name,attr,optional"`
 	CAFile                  string            `river:"ca_file,attr,optional"`
 	CertFile                string            `river:"cert_file,attr,optional"`
 	KeyFile                 string            `river:"key_file,attr,optional"`
@@ -43,11 +49,21 @@ type Arguments struct {
 	ZookeeperURIs           []string          `river:"zookeeper_uris,attr,optional"`
 	ClusterName             string            `river:"kafka_cluster_name,attr,optional"`
 	MetadataRefreshInterval string            `river:"metadata_refresh_interval,attr,optional"`
+	ServiceName             string            `river:"gssapi_service_name,attr,optional"`
+	KerberosConfigPath      string            `river:"gssapi_kerberos_config_path,attr,optional"`
+	Realm                   string            `river:"gssapi_realm,attr,optional"`
+	KeyTabPath              string            `river:"gssapi_key_tab_path,attr,optional"`
+	KerberosAuthType        string            `river:"gssapi_kerberos_auth_type,attr,optional"`
+	OffsetShowAll           bool              `river:"offset_show_all,attr,optional"`
+	TopicWorkers            int               `river:"topic_workers,attr,optional"`
 	AllowConcurrent         bool              `river:"allow_concurrency,attr,optional"`
+	AllowAutoTopicCreation  bool              `river:"allow_auto_topic_creation,attr,optional"`
 	MaxOffsets              int               `river:"max_offsets,attr,optional"`
-	PruneIntervalSeconds    int               `river:"prune_interval_seconds,attr,optional"`
+	PruneIntervalSeconds    int               `river:"prune_interval_seconds,attr,optional"` // deprecated - no-op
 	TopicsFilter            string            `river:"topics_filter_regex,attr,optional"`
+	TopicsExclude           string            `river:"topics_exclude_regex,attr,optional"`
 	GroupFilter             string            `river:"groups_filter_regex,attr,optional"`
+	GroupExclude            string            `river:"groups_exclude_regex,attr,optional"`
 }
 
 func init() {
@@ -99,7 +115,9 @@ func (a *Arguments) Convert() *kafka_exporter.Config {
 		SASLUsername:            a.SASLUsername,
 		SASLPassword:            config.Secret(a.SASLPassword),
 		SASLMechanism:           a.SASLMechanism,
+		SASLDisablePAFXFast:     a.SASLDisablePAFXFast,
 		UseTLS:                  a.UseTLS,
+		TlsServerName:           a.TlsServerName,
 		CAFile:                  a.CAFile,
 		CertFile:                a.CertFile,
 		KeyFile:                 a.KeyFile,
@@ -109,10 +127,20 @@ func (a *Arguments) Convert() *kafka_exporter.Config {
 		ZookeeperURIs:           a.ZookeeperURIs,
 		ClusterName:             a.ClusterName,
 		MetadataRefreshInterval: a.MetadataRefreshInterval,
+		ServiceName:             a.ServiceName,
+		KerberosConfigPath:      a.KerberosConfigPath,
+		Realm:                   a.Realm,
+		KeyTabPath:              a.KeyTabPath,
+		KerberosAuthType:        a.KerberosAuthType,
+		OffsetShowAll:           a.OffsetShowAll,
+		TopicWorkers:            a.TopicWorkers,
 		AllowConcurrent:         a.AllowConcurrent,
+		AllowAutoTopicCreation:  a.AllowAutoTopicCreation,
 		MaxOffsets:              a.MaxOffsets,
 		PruneIntervalSeconds:    a.PruneIntervalSeconds,
 		TopicsFilter:            a.TopicsFilter,
+		TopicsExclude:           a.TopicsExclude,
 		GroupFilter:             a.GroupFilter,
+		GroupExclude:            a.GroupExclude,
 	}
 }